bypass NotFound errors from css and js files
This has been implemented for the errors we get while running buildout.
The actual problems come from core Plone becuase of changes to the
registry in Plone 5. This will bypass the errors so we don't get a PDB
prompt during local development, but still put a message in the log.